9.0 imp check (#28)
* IMP

* IMP

* REF

* FIX

* IMP

+++ account_check/models/account_check.py
@@ -27,16 +27,20 @@ class AccountCheckOperation(models.Model):
         ondelete='cascade'
     )
     operation = fields.Selection([
-        # ('draft', 'Draft'),
-        ('holding', 'Payment'),
+        # from payments
+        ('holding', 'Receive'),
         ('deposited', 'Deposit'),
-        ('endorsed', 'Endorsement'),
+        ('selled', 'Sell'),
+        ('delivered', 'Deliver'),
         ('handed', 'Hand'),
+        ('withdrawed', 'Withdrawal'),
+        # from checks
+        ('reclaimed', 'Claim'),
         ('rejected', 'Rejection'),
         ('debited', 'Debit'),
         ('returned', 'Return'),
         ('changed', 'Change'),
-        # ('cancel', 'Cancel'),
+        ('cancel', 'Cancel'),
     ],
         required=True,
     )
@@ -44,6 +48,9 @@ class AccountCheckOperation(models.Model):
     #     'account.move.line',
     #     ondelete='cascade',
     # )
+    origin_name = fields.Char(
+        compute='_compute_origin_name'
+    )
     origin = fields.Reference(
         string='Origin Document',
         selection='_reference_models')
@@ -51,11 +58,53 @@ class AccountCheckOperation(models.Model):
         'res.partner',
         string='Partner',
     )
+    notes = fields.Text(
+    )
+
+    @api.multi
+    def unlink(self):
+        for rec in self:
+            if rec.origin:
+                raise ValidationError(_(
+                    'You can not delete a check operation that has an origin.'
+                    '\nYou can delete the origin reference and unlink after.'))
+        return super(AccountCheckOperation, self).unlink()
+
+    # no longer needed because we try to autoclean origin
+    # @api.multi
+    # def clean_origin(self):
+    #     self.origin = False
+
+    @api.multi
+    @api.depends('origin')
+    def _compute_origin_name(self):
+        """
+        We add this computed method because an error on tree view displaying
+        reference field when destiny record is deleted.
+        As said in this post (last answer) we should use name_get instead of
+        display_name
+        https://www.odoo.com/es_ES/forum/ayuda-1/question/
+        how-to-override-name-get-method-in-new-api-61228
+        """
+        for rec in self:
+            try:
+                if rec.origin:
+                    id, name = rec.origin.name_get()[0]
+                    origin_name = name
+                    # origin_name = rec.origin.display_name
+                else:
+                    origin_name = False
+            except:
+                # if we can get origin we clean it
+                rec.write({'origin': False})
+                origin_name = False
+            rec.origin_name = origin_name
 
     @api.model
     def _reference_models(self):
         return [
-            ('account.check', 'Payment'),
+            ('account.payment', 'Payment'),
+            ('account.check', 'Check'),
             ('account.invoice', 'Invoice'),
             ('account.move', 'Journal Entry'),
             ('account.move.line', 'Journal Item'),
@@ -107,7 +156,10 @@ class AccountCheck(models.Model):
         ('draft', 'Draft'),
         ('holding', 'Holding'),
         ('deposited', 'Deposited'),
-        ('endorsed', 'Endorsed'),
+        ('selled', 'Selled'),
+        ('delivered', 'Delivered'),
+        ('reclaimed', 'Reclaimed'),
+        ('withdrawed', 'Withdrawed'),
         ('handed', 'Handed'),
         ('rejected', 'Rejected'),
         ('debited', 'Debited'),
@@ -116,7 +168,8 @@ class AccountCheck(models.Model):
         ('cancel', 'Cancel'),
     ],
         required=True,
-        track_visibility='onchange',
+        # no need, operations are the track
+        # track_visibility='onchange',
         default='draft',
         copy=False,
         compute='_compute_state',
@@ -295,9 +348,23 @@ class AccountCheck(models.Model):
         return True
 
     @api.multi
-    def _del_operation(self):
+    def _del_operation(self, operation):
+        """
+        We check that the operation that is being cancel is the last operation
+        done (same as check state)
+        """
         for rec in self:
+            if operation and rec.state != operation:
+                raise ValidationError(_(
+                    'You can not cancel operation "%s" if check is in '
+                    '"%s" state') % (
+                        rec.operation_ids._fields[
+                            'operation'].convert_to_export(
+                                operation, rec.env),
+                        rec._fields['state'].convert_to_export(
+                            rec.state, rec.env)))
             if rec.operation_ids:
+                rec.operation_ids[0].origin = False
                 rec.operation_ids[0].unlink()
 
     @api.multi
@@ -318,9 +385,58 @@ class AccountCheck(models.Model):
     )
     def _compute_state(self):
         for rec in self:
-            rec.state = (
-                rec.operation_ids and
-                rec.operation_ids[0].operation or 'draft')
+            if rec.operation_ids:
+                operation = rec.operation_ids[0].operation
+                rec.state = operation
+            else:
+                rec.state = 'draft'
+
+    @api.multi
+    def _check_state_change(self, operation):
+        """
+        We only check state change from _add_operation because we want to
+        leave the user the possibility of making anything from interface.
+        On operation_from_state_map dictionary:
+        * key is 'to state'
+        * value is 'from states'
+        """
+        self.ensure_one()
+        # if we do it from _add_operation only, not from a contraint of before
+        # computing the value, we can just read it
+        old_state = self.state
+        # try:
+        #     old_state = self.read(['state'])[0]['state']
+        # except:
+        #     return True
+        operation_from_state_map = {
+            # 'draft': [False],
+            'holding': ['draft', 'deposited', 'selled', 'delivered'],
+            'delivered': ['holding'],
+            'deposited': ['holding', 'rejected'],
+            'selled': ['holding'],
+            'handed': ['draft'],
+            'withdrawed': ['draft'],
+            'rejected': ['delivered', 'deposited', 'selled', 'handed'],
+            'debited': ['handed'],
+            'returned': ['handed'],
+            'changed': ['handed'],
+            'cancel': ['draft'],
+            'reclaimed': ['rejected'],
+        }
+        from_states = operation_from_state_map.get(operation)
+        if not from_states:
+            raise ValidationError(_(
+                'Operation %s not implemented for checks!') % operation)
+        if old_state not in from_states:
+            raise ValidationError(_(
+                'You can not "%s" a check from state "%s"!\n'
+                'Check nbr (id): %s (%s)') % (
+                    self.operation_ids._fields[
+                        'operation'].convert_to_export(
+                            operation, self.env),
+                    self._fields['state'].convert_to_export(
+                        old_state, self.env),
+                    self.name, self.id))
 
     @api.multi
     def unlink(self):
@@ -329,6 +445,195 @@ class AccountCheck(models.Model):
                 raise ValidationError(
                     _('The Check must be in draft state for unlink !'))
         return super(AccountCheck, self).unlink()
+
+# checks operations from checks
+
+    @api.multi
+    def bank_debit(self):
+        self.ensure_one()
+        if self.state in ['handed']:
+            origin = self.operation_ids[0].origin
+            if origin._name != 'account.payment':
+                raise ValidationError((
+                    'The deposit operation is not linked to a payment.'
+                    'If you want to reject you need to do it manually.'))
+            vals = self.get_bank_vals(
+                'bank_debit', origin.journal_id)
+            move = self.env['account.move'].create(vals)
+            move.post()
+            # self.env['account.move'].create({
+            # })
+            self._add_operation('debited', move)
+
+    @api.multi
+    def claim(self):
+        if self.state in ['rejected'] and self.type == 'third_check':
+            operation = self._get_operation('holding', True)
+            return self.action_create_debit_note(
+                'reclaimed', 'customer', operation.partner_id)
+
+    @api.multi
+    def _get_operation(self, operation, partner_required=False):
+        self.ensure_one()
+        operation = self.operation_ids.search([
+            ('check_id', '=', self.id), ('operation', '=', operation)],
+            limit=1)
+        if partner_required:
+            if not operation.partner_id:
+                raise ValidationError((
+                    'The %s operation has no partner linked.'
+                    'You will need to do it manually.') % operation)
+        return operation
+
+    @api.multi
+    def reject(self):
+        self.ensure_one()
+        if self.state in ['deposited', 'selled']:
+            operation = self._get_operation(self.state)
+            if operation.origin._name != 'account.payment':
+                raise ValidationError((
+                    'The deposit operation is not linked to a payment.'
+                    'If you want to reject you need to do it manually.'))
+            vals = self.get_bank_vals(
+                'bank_reject', operation.origin.destination_journal_id)
+            move = self.env['account.move'].create(vals)
+            move.post()
+            self._add_operation('rejected', move)
+        elif self.state in ['delivered', 'handed']:
+            operation = self._get_operation(self.state, True)
+            return self.action_create_debit_note(
+                'rejected', 'supplier', operation.partner_id)
+
+    @api.multi
+    def action_create_debit_note(self, operation, partner_type, partner):
+        self.ensure_one()
+
+        if partner_type == 'supplier':
+            invoice_type = 'in_invoice'
+            journal_type = 'purchase'
+            view_id = self.env.ref('account.invoice_supplier_form').id
+        else:
+            invoice_type = 'out_invoice'
+            journal_type = 'sale'
+            view_id = self.env.ref('account.invoice_form').id
+
+        journal = self.env['account.journal'].search([
+            ('company_id', '=', self.company_id.id),
+            ('type', '=', journal_type),
+        ], limit=1)
+
+        name = _('Check "%s" rejection') % (self.name)
+
+        inv_line_vals = {
+            # 'product_id': self.product_id.id,
+            'name': name,
+            'account_id': self.company_id._get_check_account('rejected').id,
+            'price_unit': (
+                self.amount_currency and self.amount_currency or self.amount),
+            # 'invoice_id': invoice.id,
+        }
+
+        inv_vals = {
+            # this is the reference that goes on account.move.line of debt line
+            # 'name': name,
+            # this is the reference that goes on account.move
+            'reference': name,
+            # 'date': self.date,
+            # 'date_invoice': self.date_invoice,
+            'origin': _('Check nbr (id): %s (%s)') % (self.name, self.id),
+            'journal_id': journal.id,
+            'partner_id': partner.id,
+            'type': invoice_type,
+            'invoice_line_ids': [(0, 0, inv_line_vals)],
+        }
+        if self.currency_id:
+            inv_vals['currency_id'] = self.currency_id.id
+        # we send internal_type for compatibility with account_document
+        invoice = self.env['account.invoice'].with_context(
+            internal_type='debit_note').create(inv_vals)
+        # raise Warning('sadas')
+
+        # invoice_line = self.env['account.invoice.line'].new(inv_line_vals)
+        # invoice_line._onchange_product_id()
+        # line_values = invoice_line._convert_to_write(invoice_line._cache)
+        # line_values['price_unit'] = self.amount
+        # invoice.write({'invoice_line_ids': [(0, 0, inv_line_vals)]})
+        # invoice.signal_workflow('invoice_open')
+        # payment_group.to_pay_move_line_ids += invoice.open_move_line_ids
+        self._add_operation(operation, invoice, partner)
+
+        return {
+            'name': name,
+            'view_type': 'form',
+            'view_mode': 'form',
+            'res_model': 'account.invoice',
+            'view_id': view_id,
+            'res_id': invoice.id,
+            'type': 'ir.actions.act_window',
+            # 'context': {
+            # #     'default_partner_id': self.partner_id.id,
+            # #     'default_company_id': self.company_id.id,
+            #     'default_type': invoice_type,
+            #     'internal_type': 'debit_note',
+            # },
+            # 'domain': [('payment_id', 'in', self.payment_ids.ids)],
+        }
+
+    @api.multi
+    def get_bank_vals(self, action, journal):
+        self.ensure_one()
+        # TODO improove how we get vals, get them in other functions
+        if action == 'bank_debit':
+            # ref = _('Debit Check Nr. ')
+            # al pagar con banco se usa esta
+            # self.journal_id.default_debit_account_id.id, al debitar
+            # tenemos que usar esa misma
+            credit_account = journal.default_debit_account_id
+            # la contrapartida es la cuenta que reemplazamos en el pago
+            debit_account = self.company_id.deferred_check_account_id
+        elif action == 'bank_reject':
+            # al transferir a un banco se usa esta. al volver tiene que volver
+            # por la opuesta
+            # self.destination_journal_id.default_credit_account_id
+            credit_account = journal.default_debit_account_id
+            debit_account = self.company_id.rejected_check_account_id
+            # credit_account_id = vou_journal.default_credit_account_id.id
+        else:
+            raise ValidationError(_(
+                'Action %s not implemented for checks!') % action)
+
+        # name = self.env['ir.sequence'].next_by_id(
+        #     journal.sequence_id.id)
+        # ref = self.name
+        name = _('Check "%s" rejection') % (self.name)
+
+        debit_line_vals = {
+            'name': name,
+            'account_id': debit_account.id,
+            # 'partner_id': partner,
+            'debit': self.amount,
+            'amount_currency': self.amount_currency,
+            'currency_id': self.currency_id.id,
+            # 'ref': ref,
+        }
+        credit_line_vals = {
+            'name': name,
+            'account_id': credit_account.id,
+            # 'partner_id': partner,
+            'credit': self.amount,
+            'amount_currency': self.amount_currency,
+            'currency_id': self.currency_id.id,
+            # 'ref': ref,
+        }
+        return {
+            'ref': name,
+            'journal_id': journal.id,
+            'date': fields.Date.today(),
+            'line_ids': [
+                (0, False, debit_line_vals),
+                (0, False, credit_line_vals)],
+            # 'ref': ref,
+        }
 
     # @api.one
     # @api.onchange('checkbook_id')
